Use map of registry addresses
