+++ .env.example
@@ -1,0 +1,1 @@
+REACT_APP_REGISTRY_ADDRESS='{"12648430":"0xe78a0f7e598cc8b0bb87894b0f60dd2a88d6a8ab"}'
