+++ .gitignore
@@ -22,3 +22,4 @@ yarn-debug.log*
 yarn-error.log*
 
 public/contracts/
+.env
