+++ src/components/stepFour/index.js
@@ -1,6 +1,11 @@
 import React from 'react'
 import '../../assets/stylesheets/application.css'
-import { checkWeb3, deployContract, getNetworkVersion } from '../../utils/blockchainHelpers'
+import {
+  checkWeb3,
+  deployContract,
+  getNetworkVersion,
+  registerCrowdsaleAddress
+} from '../../utils/blockchainHelpers'
 import {
   addWhiteListRecursive,
   download,
@@ -101,7 +106,7 @@ const { PUBLISH } = NAVIGATION_STEPS
         .then(() => this.deployToken())
         .then(() => this.deployPricingStrategy())
         .then(() => this.deployCrowdsale())
-        .then(() => this.registerCrowdsaleAddress(web3))
+        .then(() => registerCrowdsaleAddress(web3, this.props.contractStore))
         .then(() => this.calculateABIEncodedArgumentsForFinalizeAgentContractDeployment())
         .then(() => this.deployFinalizeAgent())
         .catch(error => this.handleError(error))
@@ -375,26 +380,6 @@ const { PUBLISH } = NAVIGATION_STEPS
             })
 
         }, Promise.resolve())
-      })
-  }
-
-  registerCrowdsaleAddress = (web3) => {
-    const toJS = x => JSON.parse(JSON.stringify(x))
-
-    const registryAbi = this.props.contractStore.registry.abi
-    const registryAddress = this.props.contractStore.registry.addr
-    const registry = new web3.eth.Contract(toJS(registryAbi), registryAddress)
-
-    return Promise.resolve()
-      .then(() => web3.eth.getAccounts())
-      .then((accounts) => accounts[0])
-      .then((account) => {
-        const crowdsaleAddress = this.props.contractStore.crowdsale.addr[0]
-        return registry.methods
-          .add('crowdsale', crowdsaleAddress, '{}')
-          .send({
-            from: account
-          })
       })
   }
