+++ src/index.js
@@ -7,8 +7,8 @@ import { Provider } from 'mobx-react';
 import * as stores from './stores';
 useStrict(true);
 
-if (!process.env['REACT_APP_REGISTRY_CONTRACT_ADDRESS']) {
-  throw new Error('REACT_APP_REGISTRY_CONTRACT_ADDRESS env variable is not present')
+if (!process.env['REACT_APP_REGISTRY_ADDRESS']) {
+  throw new Error('REACT_APP_REGISTRY_ADDRESS env variable is not present')
 }
 
 ReactDOM.render(
